Renamed com.ncc.gaia to com.ncc.aif.
