+++ src/main/java/com/ncc/aif/OntologyMapping.java
@@ -1,4 +1,4 @@
-package com.ncc.gaia;
+package com.ncc.aif;
 
 import org.apache.jena.rdf.model.Resource;
